edit Detail

+++ src/app/shared/component/article-detail/comment/comment.component.ts
@@ -57,6 +57,7 @@ export class CommentComponent implements OnInit {
         .subscribe(
           (data) => {
             this.comments.unshift(data.comment);
+            (document.getElementById('InputComment') as HTMLFormElement).value='';
           },
           (err) => {
             console.log(err);
@@ -66,29 +67,49 @@ export class CommentComponent implements OnInit {
   }
 
   deleteComment(commentId: number) {
-    this.commentService.deleteArticleComment(this.slug, commentId).subscribe(
-      () => {
-        const Toast = Swal.mixin({
-          toast: true,
-          position: 'top-end',
-          showConfirmButton: false,
-          timer: 3000,
-          timerProgressBar: true,
-          didOpen: (toast) => {
-            toast.addEventListener('mouseenter', Swal.stopTimer);
-            toast.addEventListener('mouseleave', Swal.resumeTimer);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#273043',
+      cancelButtonColor: '#DC3545',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.commentService.deleteArticleComment(this.slug, commentId).subscribe(
+          () => {
+            const Toast = Swal.mixin({
+              toast: true,
+              position: 'top-end',
+              showConfirmButton: false,
+              timer: 3000,
+              timerProgressBar: true,
+              didOpen: (toast) => {
+                toast.addEventListener('mouseenter', Swal.stopTimer);
+                toast.addEventListener('mouseleave', Swal.resumeTimer);
+              },
+            });
+    
+            Toast.fire({
+              icon: 'success',
+              title: 'Delete comment successfully',
+            });
+            this.comments.splice(this.findComment(commentId), 1);
           },
-        });
-
-        Toast.fire({
+          (err) => {
+            console.log(err);
+          }
+        );
+        Swal.fire({
+          title: 'Deleted!',
+          text: 'Your file has been deleted.',
           icon: 'success',
-          title: 'Delete comment successfully',
+          showCancelButton: false,
+          confirmButtonColor: '#273043',
         });
-        this.comments.splice(this.findComment(commentId), 1);
-      },
-      (err) => {
-        console.log(err);
       }
-    );
+    });
+    
   }
 }
